Re-organize caching code to accommodate non-default caching strategies.
