+++ src/sequence/cache/sequence-cache.js
@@ -1,13 +1,7 @@
-const { assign, omit, values } = require("lodash");
+const { values } = require("lodash");
 
-const {
-  BaseSequence,
-  sequence: makeNewSequence,
-  EXHAUSTED
-} = require("./sequence");
-const compress = require("./compress");
-
-const cache = Object.create(null);
+const { EXHAUSTED, BaseSequence } = require("../sequence");
+const compress = require("../compress");
 
 
 /**
@@ -21,15 +15,20 @@ class ForkedSequence extends BaseSequence {
   }
 }
 
+
+function noop () {}
+
+
 /**
  * An object that forks an input source sequence into multiple output sequences.
  */
 class SequenceCache {
-  constructor (source) {
+  constructor (source, onCompress) {
     this.source = source;
     this.sourceCursor = 0;
     this.buffer = [];
     this.compressedBuffer = null;
+    this.onCompress = onCompress || noop;
 
     this.forks = 0;
     this.forkCursors = Object.create(null);
@@ -132,45 +131,11 @@ class SequenceCache {
     this.sourceCursor = null;
     this.forks = null;
     this.forkCursors = null;
+    this.onCompress(this);
   }
 }
 
-/**
- * Checks whether a node is cacheable.
- *
- * If it is not cachable, return the sequence that would've been created without
- * caching being involved.
- *
- * If the node _is_ cacheable, return either a fork of a previously-cached
- * sequence (with the same key), or a new fork if an identical node was not
- * previously cached.
- *
- * @param      {Sequence}   sequence         The non-cached sequence.
- * @param      {VDOM}       node             VDOM node to render and/or cache.
- * @param      {Function}   sequenceFactory  Function that returns a sequence for the node.
- *
- * @return     {Sequence}    The non-cached or cached sequence.
- */
-function getCachedSequence (sequence, node, sequenceFactory) {
-  const cacheKey = node.props && node.props.cacheKey;
-  if (!cacheKey) {
-    return sequenceFactory(sequence, node);
-  }
 
-  let cacheEntry = cache[cacheKey];
-
-  if (!cacheEntry) {
-    const _node = assign({}, node, {
-      props: omit(node.props, ["cacheKey"])
-    });
-
-    const sequenceToCache = makeNewSequence();
-    sequenceFactory(sequenceToCache, _node);
-    cacheEntry = cache[cacheKey] = new SequenceCache(sequenceToCache);
-  }
-
-  return cacheEntry.fork();
-}
-
-
-module.exports = getCachedSequence;
+module.exports = {
+  SequenceCache
+};
